[AN] feat(#552): 마이스터디 메인 화면 로직 수정
activity -> fragment
