--- android/gitudy/presentation/src/main/java/com/takseha/presentation/ui/mystudy/ToDoActivity.kt
@@ -1,12 +1,0 @@
-package com.takseha.presentation.ui.mystudy
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.takseha.presentation.R
-
-class ToDoActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_to_do)
-    }
-}
