+++ android/gitudy/presentation/src/main/res/layout/fragment_commit_web_view.xml
@@ -9,7 +9,7 @@
 
     <androidx.cardview.widget.CardView
         android:layout_width="match_parent"
-        android:layout_height="890dp"
+        android:layout_height="860dp"
         android:elevation="50dp"
         app:cardCornerRadius="16dp"
         android:layout_marginBottom="-20dp"
