+++ android/gitudy/presentation/src/main/res/layout/fragment_to_do.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.mystudy.ToDoFragment"
-    android:paddingBottom="30dp">
+    android:paddingBottom="30dp"
+    tools:context=".ui.mystudy.ToDoFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/todoTitle"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginLeft="22dp"
-        android:layout_marginRight="22dp"
         android:layout_marginTop="20dp"
+        android:layout_marginRight="22dp"
         app:layout_constraintTop_toTopOf="parent">
 
         <ImageView
@@ -38,10 +38,10 @@
     </androidx.constraintlayout.widget.ConstraintLayout>
 
     <CheckBox
-        android:layout_marginLeft="22dp"
         android:id="@+id/commitWithTodoCheckBtn"
         android:layout_width="22dp"
         android:layout_height="22dp"
+        android:layout_marginLeft="22dp"
         android:layout_marginTop="22dp"
         android:background="@drawable/selector_ic_check_sq_checked"
         android:button="@null"
@@ -105,12 +105,25 @@
         android:id="@+id/todoWithDateList"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_marginTop="20dp"
         android:layout_marginLeft="22dp"
+        android:layout_marginTop="20dp"
         android:layout_marginRight="22dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/memberList"
         tools:layout_editor_absoluteX="44dp" />
+
+    <ImageView
+        android:id="@+id/addTodoBtn"
+        android:layout_width="56dp"
+        android:layout_height="56dp"
+        android:layout_marginEnd="28dp"
+        android:layout_marginBottom="48dp"
+        android:background="@drawable/circle_non_stroke"
+        android:elevation="5dp"
+        android:padding="8dp"
+        android:src="@drawable/ic_add_todo"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent" />
 
     <androidx.cardview.widget.CardView
         android:id="@+id/TodoPopup"
