--- android/gitudy/presentation/src/main/res/navigation/todo_nav.xml
@@ -1,40 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/todo_nav"
-    app:startDestination="@id/toDoFragment">
-    <fragment
-        android:id="@+id/toDoFragment"
-        android:name="com.takseha.presentation.ui.mystudy.ToDoFragment"
-        android:label="fragment_to_do"
-        tools:layout="@layout/fragment_to_do" >
-        <action
-            android:id="@+id/action_toDoFragment_to_updateTodoFragment"
-            app:destination="@id/updateTodoFragment" />
-        <action
-            android:id="@+id/action_toDoFragment_to_myCommitFragment"
-            app:destination="@id/myCommitFragment"
-            app:enterAnim="@anim/vertical_enter"
-            app:exitAnim="@anim/vertical_exit" />
-    </fragment>
-    <fragment
-        android:id="@+id/updateTodoFragment"
-        android:name="com.takseha.presentation.ui.mystudy.UpdateTodoFragment"
-        android:label="fragment_update_todo"
-        tools:layout="@layout/fragment_update_todo" />
-    <fragment
-        android:id="@+id/myCommitFragment"
-        android:name="com.takseha.presentation.ui.profile.CommitDetailFragment"
-        android:label="fragment_my_commit_detail"
-        tools:layout="@layout/fragment_commit_detail" >
-        <action
-            android:id="@+id/action_myCommitFragment_to_commitWebViewFragment"
-            app:destination="@id/commitWebViewFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/commitWebViewFragment"
-        android:name="com.takseha.presentation.ui.profile.CommitWebViewFragment"
-        android:label="fragment_commit_web_view"
-        tools:layout="@layout/fragment_commit_web_view" />
-</navigation>
