+++ android/gitudy/presentation/src/main/res/values/strings.xml
@@ -46,7 +46,7 @@
     <string name="study_unlock">스터디 공개</string>
     <string name="study_deleted">스터디 종료</string>
     <string name="study_team_rank">%d점 • %d등</string>
-    <string name="study_member_score">%d점</string>
+    <string name="study_member_score">기여도 %d</string>
     <string name="study_team_rank_full">랭킹 %d등 • 최근활동 %s</string>
     <string name="study_github_link">https://github.com/%s/%s</string>
     <string name="study_noti">2월까지 프로젝트 마무리!</string>
@@ -132,6 +132,8 @@
     <string name="alert_logout_cancel">닫기</string>
     <string name="alert_logout_confirm">로그아웃</string>
     <string name="alert_delete_account_confirm">탈퇴하기</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 
 </resources>
